pyproject.toml
[tool.poetry]
name = "sweetyaar"
version = "0.1.0"
description = ""
authors = ["Zach Moshe"]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.8"
aiohttp-jinja2 = ""
jinja2 = ""
pyalsaaudio = "^0.9"
pyyaml = ""

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
